Fix unit tests for Python 3.11+
Python 3.11 changed the implementation of Enum.__format__ method, which now
returns the name of the enum (e.g.  "Catalog.SOURCE") rather than the value
("source"). This broke unit tests which relied on regex matching the value.

This patch fixes the problem in a way which keeps the compatibility with
both older and newer versions of Python.